// models/user-profile-extended.tsx
export type UserProfileExtended = {
  fullName: string;
  department: string;
  year: string;
  sleepSchedule: string;
  cleanliness: string;
  guestPolicy: string;
  sharing: string;
  socialStyle: string;
  musicTaste: string[];
  similarityPercentage: number;
  similaritySummary: string;
  id?: string;
};